+++ radio/src/translations/tts_es.cpp
@@ -190,7 +190,7 @@ I18N_PLAY_FUNCTION(es, playNumber, getvalue_t number, uint8_t unit, uint8_t att)
   }
 }
 
-I18N_PLAY_FUNCTION(es, playDuration, int16_t seconds PLAY_DURATION_ATT)
+I18N_PLAY_FUNCTION(es, playDuration, int seconds PLAY_DURATION_ATT)
 {
   if (seconds < 0) {
     PUSH_NUMBER_PROMPT(ES_PROMPT_MENO);
